literature/requirements.txt
playwright>=1.40.0
requests>=2.31.0
PyYAML>=6.0.1
beautifulsoup4>=4.12.0
scholarly>=1.7.11
